rename richtext editor
