+++ src/common/components/RichTextEditor/components/Toolbar/styled.tsx
