+++ src/common/components/RichTextEditor/configs/autoformatRules.tsx
