+++ src/common/components/RichTextEditor/configs/resetNodeOptions.tsx
