+++ src/common/components/RichTextEditor/configs/softExitOptions.tsx
