+++ src/common/components/RichTextEditor/index.tsx
@@ -8,9 +8,9 @@ import { HighlightHTML } from '../HighlightHTML'
 import { Toolbar } from './components/Toolbar'
 import { INITIAL_CONTENT } from './constants'
 import { plugins } from './plugins'
-import { RichTextEditorV2Props } from './types'
+import { RichTextEditorProps } from './types'
 
-export const RichTextEditorV2: React.FC<RichTextEditorV2Props> = ({ id, defaultValue = INITIAL_CONTENT, onChange }) => {
+export const RichTextEditor: React.FC<RichTextEditorProps> = ({ id, defaultValue = INITIAL_CONTENT, onChange }) => {
   const theme = useTheme()
 
   const editableProps: EditableProps = {
