+++ src/common/components/RichTextEditor/types.ts
@@ -1,6 +1,6 @@
 import { TNode } from '@udecode/plate-core'
 
-export interface RichTextEditorV2Props {
+export interface RichTextEditorProps {
   id: string
   defaultValue?: TNode[]
   onChange?: (newValue: TNode[]) => void
