+++ src/modules/CourseDetail/context/Review/index.tsx
@@ -6,8 +6,8 @@ import React, { createContext, useContext, useEffect } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import toast from 'react-hot-toast'
 
-import { INITIAL_CONTENT } from '@/common/components/RichTextV2/constants'
-import { plugins } from '@/common/components/RichTextV2/plugins'
+import { INITIAL_CONTENT } from '@/common/components/RichTextEditor/constants'
+import { plugins } from '@/common/components/RichTextEditor/plugins'
 import { useCourseGroup } from '@/common/hooks/useCourseGroup'
 import { Storage } from '@/common/storage'
 import { StorageKey } from '@/common/storage/constants'
